Moved enum classes to own folder
