+++ SpringServer/src/main/java/com/afa/enums/FamilySize.java
@@ -1,3 +1,3 @@
-package com.afa.model;
+package com.afa.enums;
 
 public enum FamilySize { SMALL, MEDIUM, LARGE };
