+++ SpringServer/src/test/java/com/afa/springserver/repositories/UserFamilyRelationshipRepositoryTests.java
@@ -1,5 +1,7 @@
 package com.afa.springserver.repositories;
 
+import com.afa.enums.FamilyRole;
+import com.afa.enums.FamilySize;
 import com.afa.model.*;
 import com.afa.repositories.*;
 import org.junit.jupiter.api.AfterEach;
