+++ SpringServer/src/test/java/com/afa/springserver/repositories/UserRepositoryTests.java
@@ -1,6 +1,6 @@
 package com.afa.springserver.repositories;
 
-import com.afa.model.FamilySize;
+import com.afa.enums.FamilySize;
 import com.afa.model.UserEntity;
 import com.afa.repositories.UserRepository;
 import org.junit.jupiter.api.Test;
